fix build error by isolating framer motion
