+++ .eslintrc.json
@@ -7,6 +7,13 @@
   },
   "rules": {
     "@typescript-eslint/no-unused-vars": ["warn"],
-    "react/react-in-jsx-scope": "off"
+    "react/react-in-jsx-scope": "off",
+    "no-restricted-syntax": [
+      "error",
+      {
+        "selector": "ExportAllDeclaration",
+        "message": "Avoid using `export *` – it's not supported in Next.js Client Components."
+      }
+    ]
   }
 }
