+++ app/forgot-password/metadata.ts
@@ -1,0 +1,17 @@
+import type { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Recover Password – Atendor',
+  description: 'Reset your Atendor password.',
+  openGraph: {
+    title: 'Recover Password – Atendor',
+    description: 'Reset your Atendor password.',
+    images: ['/og.png'],
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Recover Password – Atendor',
+    description: 'Reset your Atendor password.',
+    images: ['/og.png'],
+  },
+}
