+++ app/login/metadata.ts
@@ -1,0 +1,17 @@
+import type { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Login – Atendor',
+  description: 'Access your Atendor account.',
+  openGraph: {
+    title: 'Login – Atendor',
+    description: 'Access your Atendor account.',
+    images: ['/og.png'],
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Login – Atendor',
+    description: 'Access your Atendor account.',
+    images: ['/og.png'],
+  },
+}
