+++ app/page.tsx
@@ -1,8 +1,7 @@
-import Image from 'next/image'
 import Link from 'next/link'
 import PublicHeader from '@/components/PublicHeader'
 import PublicFooter from '@/components/PublicFooter'
-import { motion } from 'framer-motion'
+import { AnimatedHero } from '@/components/AnimatedHero'
 
 export const metadata = {
   title: 'Atendor – AI Visual Builder',
@@ -25,30 +24,7 @@ export default function Home() {
     <main className="space-y-24">
       <PublicHeader />
       {/* Hero */}
-      <section className="relative flex flex-col items-center justify-center min-h-screen px-4 text-center bg-gradient-to-b from-white via-slate-50 to-slate-100">
-        <motion.div
-          initial={{ opacity: 0, y: 40 }}
-          whileInView={{ opacity: 1, y: 0 }}
-          viewport={{ once: true }}
-          transition={{ duration: 0.6 }}
-          className="space-y-6"
-        >
-          <h1 className="text-5xl font-bold tracking-tight md:text-6xl">
-            Build AI Chat Assistants Visually
-          </h1>
-          <p className="mx-auto max-w-xl text-gray-600 md:text-xl">
-            No code. No complexity. Just powerful conversations connected to your tools.
-          </p>
-          <div className="flex flex-col items-center justify-center gap-4 sm:flex-row">
-            <Link href="/signup" className="px-6 py-3 text-white bg-black rounded-md hover:bg-slate-800 transition-colors">
-              Get Started
-            </Link>
-            <Link href="/login" className="px-6 py-3 border rounded-md hover:bg-slate-100 transition-colors">
-              Login
-            </Link>
-          </div>
-        </motion.div>
-      </section>
+      <AnimatedHero />
 
       {/* How It Works */}
       <section className="px-4 py-16 mx-auto max-w-5xl space-y-12">
@@ -78,46 +54,33 @@ export default function Home() {
         <div className="mx-auto max-w-5xl space-y-12">
           <h2 className="text-3xl font-semibold text-center">Key Features</h2>
           <div className="grid gap-8 md:grid-cols-4">
-            <motion.div
-              whileInView={{ opacity: 1, scale: 1 }}
-              initial={{ opacity: 0, scale: 0.9 }}
-              transition={{ duration: 0.4 }}
-              viewport={{ once: true }}
-              className="space-y-2 text-center"
+            <div
+              className="space-y-2 text-center opacity-0 animate-in fade-in zoom-in-90 duration-400"
             >
               <div className="text-xl font-bold">Visual Builder</div>
               <p>No code, flexible and intuitive.</p>
-            </motion.div>
-            <motion.div
-              whileInView={{ opacity: 1, scale: 1 }}
-              initial={{ opacity: 0, scale: 0.9 }}
-              transition={{ duration: 0.4, delay: 0.1 }}
-              viewport={{ once: true }}
-              className="space-y-2 text-center"
+            </div>
+            <div
+              className="space-y-2 text-center opacity-0 animate-in fade-in zoom-in-90 duration-400"
+              style={{ animationDelay: '100ms' }}
             >
               <div className="text-xl font-bold">WhatsApp/Discord</div>
               <p>Reach users where they are.</p>
-            </motion.div>
-            <motion.div
-              whileInView={{ opacity: 1, scale: 1 }}
-              initial={{ opacity: 0, scale: 0.9 }}
-              transition={{ duration: 0.4, delay: 0.2 }}
-              viewport={{ once: true }}
-              className="space-y-2 text-center"
+            </div>
+            <div
+              className="space-y-2 text-center opacity-0 animate-in fade-in zoom-in-90 duration-400"
+              style={{ animationDelay: '200ms' }}
             >
               <div className="text-xl font-bold">Role-based Access</div>
               <p>Control who edits flows.</p>
-            </motion.div>
-            <motion.div
-              whileInView={{ opacity: 1, scale: 1 }}
-              initial={{ opacity: 0, scale: 0.9 }}
-              transition={{ duration: 0.4, delay: 0.3 }}
-              viewport={{ once: true }}
-              className="space-y-2 text-center"
+            </div>
+            <div
+              className="space-y-2 text-center opacity-0 animate-in fade-in zoom-in-90 duration-400"
+              style={{ animationDelay: '300ms' }}
             >
               <div className="text-xl font-bold">Supabase Backend</div>
               <p>Built for performance and scale.</p>
-            </motion.div>
+            </div>
           </div>
         </div>
       </section>
@@ -134,7 +97,7 @@ export default function Home() {
       {/* Who it's for */}
       <section className="px-4 py-16 bg-slate-50">
         <div className="mx-auto max-w-5xl space-y-12">
-          <h2 className="text-3xl font-semibold text-center">Who It's For</h2>
+          <h2 className="text-3xl font-semibold text-center">Who It&apos;s For</h2>
           <div className="grid gap-8 md:grid-cols-3">
             <div className="space-y-2 text-center">
               <div className="text-xl font-bold">Solo Creators</div>
@@ -165,9 +128,9 @@ export default function Home() {
         <div className="mx-auto max-w-5xl space-y-8 text-center">
           <h2 className="text-3xl font-semibold">What Users Say</h2>
           <div className="grid gap-8 md:grid-cols-3">
-            <div className="p-4 bg-white rounded shadow">"Amazing tool!"</div>
-            <div className="p-4 bg-white rounded shadow">"Built my bot in a day."</div>
-            <div className="p-4 bg-white rounded shadow">"Our team loves it."</div>
+            <div className="p-4 bg-white rounded shadow">&quot;Amazing tool!&quot;</div>
+            <div className="p-4 bg-white rounded shadow">&quot;Built my bot in a day.&quot;</div>
+            <div className="p-4 bg-white rounded shadow">&quot;Our team loves it.&quot;</div>
           </div>
         </div>
       </section>
