+++ app/reset-password/metadata.ts
@@ -1,0 +1,17 @@
+import type { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Reset Password – Atendor',
+  description: 'Choose a new password for your account.',
+  openGraph: {
+    title: 'Reset Password – Atendor',
+    description: 'Choose a new password for your account.',
+    images: ['/og.png'],
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Reset Password – Atendor',
+    description: 'Choose a new password for your account.',
+    images: ['/og.png'],
+  },
+}
