+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@supabase/auth-helpers-nextjs": "^0.10.0",
     "@supabase/supabase-js": "^2.50.0",
+    "framer-motion": "^12.17.0",
     "lucide-react": "^0.513.0",
     "magic-ui": "^0.1.0",
     "nanoid": "^5.1.5",
